requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
Flask-Cors==4.0.0
Werkzeug==3.1.3
python-dotenv==1.0.0
pandas
openpyxl
reportlab==4.0.4
APScheduler==3.11.0
